+++ packages/snap-networks-utils/src/index.ts
@@ -1,13 +1,20 @@
 export type { Logger } from './logger';
 export { createPrefixedLogger, logger, noOpLogger } from './logger';
 export { AssetsService } from './services/assets/AssetsService';
+export {
+  ASSETS_SERVICE_NAME,
+  type AssetsServiceMessenger,
+  type AssetsServiceMessengerCaller,
+} from './services/assets/messenger';
 export type { AssetEntity, AssetScope } from './services/assets/types';
 export { mapControllerAsset } from './services/assets/utils/mapControllerAsset';
-export type {
-  AssetsControllerGetAssetAction,
-  AssetsControllerGetAssetsAction,
-  CoreMessenger,
-  CoreMessengerActions,
-  CoreMessengerCaller,
+export {
+  CORE_MESSENGER_NAMESPACE,
+  type AssetsControllerGetAccountAssetByIDAction,
+  type AssetsControllerGetAccountAssetsByIDsAction,
+  type AssetsControllerGetAccountAssetsByScopeAction,
+  type CoreMessenger,
+  type CoreMessengerActions,
+  type CoreMessengerCaller,
 } from './types/core-messenger';
 export { toUiAmount } from './utils/toUiAmount';